+++ src/oracle_implementations/spot_price/Chainlink/ChainLinkValueProvider.t.sol
@@ -47,10 +47,7 @@ contract ChainLinkValueProviderTest is DSTest {
             abi.encodeWithSelector(
                 IChainlinkAggregatorV3Interface.decimals.selector
             ),
-            MockProvider.ReturnData({
-                success: true,
-                data: abi.encode(uint8(8))
-            }),
+            MockProvider.ReturnData({success: true, data: abi.encode(8)}),
             false
         );
